fix: restore the Equals/GetHashCode contract on B2Vec2
Moving operator== to an IEEE compare was the right call, it matches C.
But Equals just forwarded to the operator, so a NaN vector stopped being
equal to itself and stopped agreeing with GetHashCode. That breaks
Dictionary<B2Vec2, _>. Keep the operator on IEEE semantics and compare
bitwise in Equals.

While here:

- b2World_Draw takes the debug draw by in. C passes b2DebugDraw*, and
  the struct has 27 fields, so it was copied every frame.
- b2GrowBitSet guards its copy the way b2CreateBitSet already guards the
  fill. Growing a bit set created with zero capacity threw, because
  b2Alloc hands back null there.
- b2ClearBit takes an int like the other three, so the call sites drop
  their casts.
- b2InternalAssert is public. It is B2_API upstream.

+++ src/Box2D.NET/B2ConstraintGraphs.cs
@@ -207,8 +207,8 @@ namespace Box2D.NET
             if (colorIndex != B2_OVERFLOW_INDEX)
             {
                 // This might clear a bit for a kinematic or static body, but this has no effect
-                b2ClearBit(ref color.bodySet, (uint)bodyIdA);
-                b2ClearBit(ref color.bodySet, (uint)bodyIdB);
+                b2ClearBit(ref color.bodySet, bodyIdA);
+                b2ClearBit(ref color.bodySet, bodyIdB);
             }
 
             int movedIndex = b2Array_RemoveSwap(ref color.contactSims, localIndex);
@@ -324,8 +324,8 @@ namespace Box2D.NET
             if (colorIndex != B2_OVERFLOW_INDEX)
             {
                 // May clear static bodies, no effect
-                b2ClearBit(ref color.bodySet, (uint)bodyIdA);
-                b2ClearBit(ref color.bodySet, (uint)bodyIdB);
+                b2ClearBit(ref color.bodySet, bodyIdA);
+                b2ClearBit(ref color.bodySet, bodyIdB);
             }
 
             int movedIndex = b2Array_RemoveSwap(ref color.jointSims, localIndex);
